backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
ultralytics
pillow
python-dotenv
httpx
websockets
slowapi
